Added a basic select function to FizzLang
This adds a very basic select method that works like the other
array manipulation methods in the module. It's very basic and in
the future may have an option to also get the integer that was
tested so more information can be found.

On branch master
Changes to be committed:
	modified:   fizzlang.rb