Refactored FizzLang.select to remove unneeded block.call
The block being passed as a parameter does not need to be wrapped
in an addiional block in order to be used with the enum select
method. Instead tihs has been substitued with passing the block
parameter to select with the block parameter syntax and sigil.

On branch master
Changes to be committed:
	modified:   fizzlang.rb