Added a with_val method to create an enumerator of results
Adding a with_val method to the FizzLang module allows the ability
to use blocks and enumerable methods to get both the value checked
and the resulting string. This helps to make enumeration methods
like select to be more useful as they have additional information.

On branch master
Changes to be committed:
	modified:   fizzlang.rb