Added additional test/example class to start.rb
This new example is just to demonstrate how the value methods are
dynamically generated and don't require you to use the terms Fizz
and Buzz. This example will instead produce strings with the
replacement of 'Foo' and 'Bar' instead of 'Fizz' and 'Buzz'
respectively. This demonstrates the power and convenience of using
modules to include and extend methods onto a class in order to
dynamically generate the class itself.

On branch master
Changes to be committed:
	modified:   start.rb