Renamed start.rb to examples.rb
The file name is clearer. The original name was based on the fact
that for testing in irb it was "started" with that file for testing.
Since now it serves as example classes it seemed appropriate to
rename it to something more clear.

On branch master
Changes to be committed:
    renamed:    start.rb -> examples.rb