Updated .gitignore to ignore any test files
This addition will ignore any file that begins with 'test'. This
was done so that test scripts don't have to end with the sh extension
in order to still be ignored by git. These are not intended to be
unit test files but rather scripts to launch irb with some default
settings to manually test with irb.

On branch master
Changes to be committed:
	modified:   .gitignore