Added some basic unit tests
Current this just tests the very basics of the FizzBuzz class
specifically. These are by no means complete unit tests and are
simply to serve as some basic checks as well as give a bit of a
demonstration of the use of the constructed classes in practice.

On branch master
Changes to be committed:
	new file:   examples_unit.rb