Refactored FizzLang.map to eliminate unneeded block nesting
The previous method used block.call inside of a map rather than
just using the shorter and simpler method of passing the block
directly to map. This change also keeps the method consistent with
the other methods that pass blocks provided as arguments.

On branch master
Changes to be committed:
	modified:   fizzlang.rb