Updated .gitignore to include new documentation directories
The local script used to generate both HTML docs and ri docs puts
the documentation in the two added folders. As stated in previous
commit, these docs are not always going to be up to date and should
be ignored in favor of running rdoc as needed.

On branch master
Changes to be committed:
	modified:   .gitignore