Added code examples to FizzLang public methods
To make usages of each method a little more clear some very basic
code examples were added to the documentation comments so that
they appear in the generated RDocs. These are not very complete at
the moment but they serve as a good starting place to later add
better examples and documentation.

On branch master
Changes to be committed:
	modified:   fizzlang.rb