Added a class method to check a range of numbers at a class level
This is just a shortcut over creating a new instance of a class
and then getting the array of results from that. The functionality
it adds simply wraps the creation of a new FizzBuzz instance. Also
adds a basic unit test for this method.

On branch master
Changes to be committed:
	modified:   examples_unit.rb
	modified:   fizzlang.rb