Added a sample showing additional ways to use class methods
This new example rather than using the class methods inside the
definition of the class calls them outside of the class as methods.
This might help show those not familiar with class methods their
similarity to static methods in other languages and also demonstrate
another way you may want to construct your classes.

On branch master
Changes to be committed:
	modified:   examples.rb