Modified Fizzy.test_cases to allow blocks to be passed
This allows a block to be passed in place of an array of args,
which for some people may make the code for creating classes a little
cleaner and more obvious what's happening. This required adding a
private method to actually handle generating the appropriate values
and calling that from test_cases instead so that it may also be
called from the process_block function to acheive the same result.

On branch master
Changes to be committed:
	modified:   fizzlang.rb