Added an example to show the new block initialization feature
Creates a simple new class called BlockBuzz in the examples that
demonstrates the new block features of calling test_cases. This
commit also adds unit tests for that new class that were copied
from FizzBuzz and modified to match BlockBuzz to ensure that classes
created this way still work as expected.

On branch master
Changes to be committed:
	modified:   examples.rb
	modified:   examples_unit.rb