Added the ability to pass keyword arguments to Fizzy.test_cases
This addition allows you to pass keyword arguments and their values
directly to test_cases rather than using the class methods or the
block method of creation. In order to facilitate this, a new private
method called process_kwargs was added to Fizzy. Since this new method
contained almost identical code to what was being employed in
process_block the latter method was modified to instead call the
former to remove duplicated code.

On branch master
Changes to be committed:
	modified:   fizzlang.rb