Refactor to allow reassignment of test conditions
Simple refactor that internally has a slight change in the way
generated methods are used in order to facilitate the ability to
reassign values to a given method. Also now adds an #{sym}= method
in order to make this a little more clear and allow reassignment
through use of the assignment operator.

On branch master
Changes to be committed:
	modified:   fizzlang.rb