Added a test class and unit tests for new assignment methods
Simple test class that reassigns a value for one of the test cases
and then unit tests to ensure the change was correctly propogated.
Interesting note, the reassignment had to be placed between comments
to stop RDoc from documenting it because it was treating it as a
new constant on Object. Could be an a bug in RDoc or could be something
strange happening behind the scenes. Something to watch out for in
future commits.

On branch master
Changes to be committed:
	modified:   examples.rb
	modified:   examples_unit.rb