Refactored check method in Fizzy to reduce complexity
The check function originally relied on a reduce function to create
the output which was somewhat harder to read. This changes the method
to use an each and a string outside the call to each. This is also
in preparation for upcoming changes that will require modification
here that will rely on this change.

On branch master
Changes to be committed:
	modified:   fizzlang.rb