Refactored and added methods to Fizzy to allow string customization
Previously the string used to replace text in a given test was
determined solely by the symbol name. With these changes it is now
possible to optionally specify a string to use for replacement. This
is supported through a number of different ways of doing it and they
will all have examples and unit tests added to them in commits
following this one.

On branch master
Changes to be committed:
	modified:   fizzlang.rb