Added an example class showing using methods for all setup
This example and its unit tests demonstrate using the class method
forms of all the DSL features to setup a class named MethodBarBaz.
This example only shows the new string methods added to the class
methods and not the other means of doing so.

On branch master
Changes to be committed:
	modified:   examples.rb
	modified:   examples_unit.rb