Added example class for using shortened class methods
This demonstrates using the shortened form of using class methods
to set values and replacement strings in the class. This example
chooses to pass the value and string to the original class methods
rather than to the addition _string class methods. This also adds
unit tests for the new example to ensure that it matches the
previous example as intended.

On branch master
Changes to be committed:
	modified:   examples.rb
	modified:   examples_unit.rb