Added example class for the new form of block setup
Like the previous commits this builds on the original block method
of creating a class and adds the ability to specify the string to
be associated with the value. Also like previous commits this adds
unit tests for the class to ensure that the class is functioning
as expected and in line with the previous examples.

On branch master
Changes to be committed:
	modified:   examples.rb
	modified:   examples_unit.rb