Added a new example class to Fizzy documentation
All of the previous examples were based on the old versions where
you weren't able to specify the strings to be used for replacement.
This is just a new example that adds one way to do this using the
shortened method call method of setting.

On branch master
Changes to be committed:
	modified:   fizzlang.rb