Modified rakefile to ignore the rakefile during documentation
Because the rakefile does not end in .rb rdoc does not treat it as
a ruby file and thus leaves the documentation not correctly rendered
as it is treated as markdown. Preferably, the rakefile would be
documented as Ruby code which it may be modified to be later.

On branch master
Changes to be committed:
	modified:   rakefile