Added ability to pass blocks to _string methods
This allows for the passing of blocks to the generated string methods
so that value returned may be based on additional conditions. The
order of precedence for the value returned is block, string, then
finally inferred from the symbol. This commit also adds the symbol
:reset to the nil check for those same string methods which now
nils the proc and sets the string to the symbol.

On branch master
Changes to be committed:
	modified:   fizzlang.rb