Added example class and tests for new block string functionality
This adds a simple example class for using blocks to get replacement
strings as well as an entirely new set of tests used to test the
manipulation of a class after the fact including checking the initial
state, changing the string value, resetting the values, changing the
string value again, and finally adding the block back again. This
ensures that all of the ways a class can be manipulated are working
as expected.

On branch master
Changes to be committed:
	modified:   examples.rb
	modified:   examples_unit.rb