Refactore Fizzy string method to use better string setting
The previous version of this dynamically generated method used the
more cryptic self.public_call method to set the string. This was
because at the time of writing the original it was planned to do more
with the _string= methods, but other changes made those obsolete.
So, those calls to public_call can simply be replaced with setting
the string in the strings hash directly.

On branch master
Changes to be committed:
	modified:   fizzlang.rb