Refactored and moved files around for building FizzLang as a gem
This requires a major refactor because of moving the fizzlang.rb
file into the lib/ directory so that the gem works as intended.
This means that every example had to have its include changed,
assuming it's running on a system that's not running the installed
gem file. This also adds the simple gemspec and modifies the build
rake task to build them gem.

On branch master
Changes to be committed:
	modified:   .gitignore
	modified:   examples/assignbuzz.rb
	modified:   examples/blockbarbaz.rb
	modified:   examples/blockbuzz.rb
	modified:   examples/blockfizzbuzz.rb
	modified:   examples/fizz.rb
	modified:   examples/fizzbuzz.rb
	modified:   examples/fizzbuzzbang.rb
	modified:   examples/fizzbuzzfoobarbazbang.rb
	modified:   examples/foobar.rb
	modified:   examples/keybuzz.rb
	modified:   examples/keywordbarbaz.rb
	modified:   examples/methodbarbaz.rb
	modified:   examples/shorterbarbaz.rb
	modified:   examples/stringybuzz.rb
	new file:   fizzlang.gemspec
	renamed:    fizzlang.rb -> lib/fizzlang.rb
	modified:   rakefile