Total refactor of all code to meet Ruby best practices
After running rubocop on the codebase a lot of styling and formatting
issues were found that needed to be addressed. This includes changing
all indentation from 4 spaces to 2 spaces, changing all CRLF to just
LF, adding an empty LF at the end of every file, adding the frozen
string literal magic comment to every file, refactoring a few things
with array and hash declarations (particularly in tests and rakefile),
breaking fizzlang.rb up into separate files for each module, moving
the private util methods for FizzLangSingleton out to their own
file and module to be included in FizzLangSingleton to shorten the
module, adding a rubocop config file, and updating all necessary
files to adjust for these changes. This is the final step in the
major refactor mentioned in previous commits. Due to the fact that
all of the file have changed the notes on committed changes below
only cover new files.

On branch master
Changes to be committed:
	new file:   lib/fizzlang_instance.rb
	new file:   lib/fizzlang_singleton.rb
	new file:   lib/fizzlang_util.rb