Updated rakefile to use yard instead of rdoc
The VSCode Solargraph plugin relies on yard for generating its in
editor documentation. The yard gem also provides very nice and clean
documentation as well as a simple documentation server daemon. So,
the documentation generator will make use of the yard gem and in
the future the documentation comments will be rewritten to use the
yard syntax.

On branch master
Changes to be committed:
	modified:   rakefile